move mjpeg vendor info into separate file
